pyproject.toml
[tool.poetry]
name = "course-work-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
celery = "^5.4.0"
gevent = "^24.2.1"
redis = "^5.0.4"
django-celery-beat = "^2.6.0"
requests = "^2.32.1"
coverage = "^7.5.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.isort]
line_length = 119



[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.tox
    | \.venv
    | dist
  )/
)
'''
